Cast -1 to (uintn) in VGID2VFILE/VGID2SLOT/VSID2VFILE/VSID2SLOT to get rid of warnings on odin. Modified explanation of VGID and VSID. VGID/VSID should be the top 8 bits, instead of the second 8 bits. A type cast in vgp.c

+++ hdf/src/vgp.c
@@ -555,7 +555,7 @@ uint16 ref;
     uint8          *vgpack;
     uint32          len;
           
-    if (( len = Hlength(f, DFTAG_VG, (uint16) ref)) == FAIL)
+    if (( len = Hlength(f, DFTAG_VG, (uint16) ref)) == (uint32)FAIL)
         HRETURN_ERROR(DFE_INTERNAL,NULL);
 
     if(( vgpack = (uint8 *) HDgetspace(len)) == NULL)
